requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
click==8.1.7
Django==5.0.4
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.0
pytest-django==4.8.0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
sqlparse==0.5.0
uritemplate==4.1.1
